support command palette registration
