+++ tabulous/widgets/_component/__init__.py
@@ -11,6 +11,7 @@ from ._column_setting import (
     ColumnDtypeInterface,
 )
 from ._proxy import ProxyInterface
+from ._viewer import Toolbar, Console, CommandPalette
 from ._table_subset import TableSeries, TableSubset, TableLocIndexer, TableILocIndexer
 
 __all__ = [
@@ -34,4 +35,7 @@ __all__ = [
     "TableSubset",
     "TableLocIndexer",
     "TableILocIndexer",
+    "Toolbar",
+    "Console",
+    "CommandPalette",
 ]
